+++ infra/content/kXml.class.php
@@ -47,8 +47,8 @@ class kXml
 	
 	public static function transformSafePhpFunction($xslt)
 	{
-		$xslt = str_replace("'gmdate'", "'myXmlUtils::kGmdate'", $xslt);
-		return str_replace("'date'", "'myXmlUtils::kDate'", $xslt);
+		$xslt = str_replace("'gmdate'", "'dateUtils::kGmdate'", $xslt);
+		return str_replace("'date'", "'dateUtils::kDate'", $xslt);
 	}
 	
 	//check if the prop's value is valid for xml encoding.